+++ src/keystore.h
@@ -1,0 +1,10 @@
+// Copyright (c) 2009-2011 Satoshi Nakamoto & Bitcoin developers
+// Distributed under the MIT/X11 software license, see the accompanying
+// file license.txt or http://www.opensource.org/licenses/mit-license.php.
+#ifndef BITCOIN_KEYSTORE_H
+#define BITCOIN_KEYSTORE_H
+
+bool AddKey(const CKey& key);
+std::vector<unsigned char> GenerateNewKey();
+
+#endif
