+++ src/ui.cpp
@@ -3,6 +3,7 @@
 // file license.txt or http://www.opensource.org/licenses/mit-license.php.
 
 #include "headers.h"
+#include "db.h"
 #include "init.h"
 #include "strlcpy.h"
 #include <boost/filesystem/fstream.hpp>
