+++ src/ui.h
@@ -4,6 +4,9 @@
 #ifndef BITCOIN_UI_H
 #define BITCOIN_UI_H
 
+#include <boost/function.hpp>
+#include "wallet.h"
+
 DECLARE_EVENT_TYPE(wxEVT_UITHREADCALL, -1)
